taking the graphs classes out of models.py
