+++ robotmbt/visualise/networkvisualiser.py
@@ -1,4 +1,5 @@
-from robotmbt.visualise.models import AbstractGraph, StateGraph
+from robotmbt.visualise.graphs.abstractgraph import AbstractGraph
+from robotmbt.visualise.graphs.stategraph import StateGraph
 from bokeh.palettes import Spectral4
 from bokeh.models import (
     Plot, Range1d, Circle, Rect,
