+++ utest/test_visualise_models.py
@@ -1,173 +1,173 @@
 import unittest
 try:
-    from robotmbt.visualise.models import *
+    from robotmbt.visualise.graphs import *
     VISUALISE = True
 except ImportError:
     VISUALISE = False
 
 if VISUALISE:
-  class TestVisualiseModels(unittest.TestCase):
-    """
-    Contains tests for robotmbt/visualise/models.py
-    """
+    class TestVisualiseModels(unittest.TestCase):
+        """
+        Contains tests for robotmbt/visualise/models.py
+        """
 
-    """
+        """
     Class: ScenarioInfo
     """
 
-    def test_scenarioInfo_str(self):
-        si = ScenarioInfo('test')
-        self.assertEqual(si.name, 'test')
-        self.assertEqual(si.src_id, 'test')
+        def test_scenarioInfo_str(self):
+            si = ScenarioInfo('test')
+            self.assertEqual(si.name, 'test')
+            self.assertEqual(si.src_id, 'test')
 
-    def test_scenarioInfo_Scenario(self):
-        s = Scenario('test')
-        s.src_id = 0
-        si = ScenarioInfo(s)
-        self.assertEqual(si.name, 'test')
-        self.assertEqual(si.src_id, 0)
+        def test_scenarioInfo_Scenario(self):
+            s = Scenario('test')
+            s.src_id = 0
+            si = ScenarioInfo(s)
+            self.assertEqual(si.name, 'test')
+            self.assertEqual(si.src_id, 0)
 
-    """
+        """
     Class: TraceInfo
     """
 
-    def test_create_TraceInfo(self):
-        ts = TraceState(3)
-        candidates = []
-        for scenario in range(3):
-            candidates.append(ts.next_candidate())
-            ts.confirm_full_scenario(candidates[-1], str(scenario), {})
-        ti = TraceInfo.from_trace_state(trace=ts, state=ModelSpace())
+        def test_create_TraceInfo(self):
+            ts = TraceState(3)
+            candidates = []
+            for scenario in range(3):
+                candidates.append(ts.next_candidate())
+                ts.confirm_full_scenario(candidates[-1], str(scenario), {})
+            ti = TraceInfo.from_trace_state(trace=ts, state=ModelSpace())
 
-        self.assertEqual(ti.trace[0].name, str(0))
-        self.assertEqual(ti.trace[1].name, str(1))
-        self.assertEqual(ti.trace[2].name, str(2))
+            self.assertEqual(ti.trace[0].name, str(0))
+            self.assertEqual(ti.trace[1].name, str(1))
+            self.assertEqual(ti.trace[2].name, str(2))
 
-        self.assertIsNotNone(ti.state)
-        # TODO check state
+            self.assertIsNotNone(ti.state)
+            # TODO check state
 
-    """
+        """
     Class: ScenarioGraph
     """
 
-    def test_scenario_graph_init(self):
-        sg = ScenarioGraph()
-        self.assertIn('start', sg.networkx.nodes)
-        self.assertEqual(sg.networkx.nodes['start']['label'], 'start')
+        def test_scenario_graph_init(self):
+            sg = ScenarioGraph()
+            self.assertIn('start', sg.networkx.nodes)
+            self.assertEqual(sg.networkx.nodes['start']['label'], 'start')
 
-    def test_scenario_graph_ids_empty(self):
-        sg = ScenarioGraph()
-        si = ScenarioInfo('test')
-        node_id = sg._get_or_create_id(si)
-        self.assertEqual(node_id, 'node0')
+        def test_scenario_graph_ids_empty(self):
+            sg = ScenarioGraph()
+            si = ScenarioInfo('test')
+            node_id = sg._get_or_create_id(si)
+            self.assertEqual(node_id, 'node0')
 
-    def test_scenario_graph_ids_duplicate_scenario(self):
-        sg = ScenarioGraph()
-        si = ScenarioInfo('test')
-        id0 = sg._get_or_create_id(si)
-        id1 = sg._get_or_create_id(si)
-        self.assertEqual(id0, id1)
+        def test_scenario_graph_ids_duplicate_scenario(self):
+            sg = ScenarioGraph()
+            si = ScenarioInfo('test')
+            id0 = sg._get_or_create_id(si)
+            id1 = sg._get_or_create_id(si)
+            self.assertEqual(id0, id1)
 
-    def test_scenario_graph_ids_different_scenarios(self):
-        sg = ScenarioGraph()
-        si0 = ScenarioInfo('test0')
-        si1 = ScenarioInfo('test1')
-        id0 = sg._get_or_create_id(si0)
-        id1 = sg._get_or_create_id(si1)
-        self.assertEqual(id0, 'node0')
-        self.assertEqual(id1, 'node1')
+        def test_scenario_graph_ids_different_scenarios(self):
+            sg = ScenarioGraph()
+            si0 = ScenarioInfo('test0')
+            si1 = ScenarioInfo('test1')
+            id0 = sg._get_or_create_id(si0)
+            id1 = sg._get_or_create_id(si1)
+            self.assertEqual(id0, 'node0')
+            self.assertEqual(id1, 'node1')
 
-    def test_scenario_graph_add_new_node(self):
-        sg = ScenarioGraph()
-        sg.ids['test'] = ScenarioInfo('test')
-        sg._add_node('test')
-        self.assertIn('test', sg.networkx.nodes)
-        self.assertEqual(sg.networkx.nodes['test']['label'], 'test')
+        def test_scenario_graph_add_new_node(self):
+            sg = ScenarioGraph()
+            sg.ids['test'] = ScenarioInfo('test')
+            sg._add_node('test')
+            self.assertIn('test', sg.networkx.nodes)
+            self.assertEqual(sg.networkx.nodes['test']['label'], 'test')
 
-    def test_scenario_graph_add_existing_node(self):
-        sg = ScenarioGraph()
-        sg._add_node('start')
-        self.assertIn('start', sg.networkx.nodes)
-        self.assertEqual(len(sg.networkx.nodes), 1)
+        def test_scenario_graph_add_existing_node(self):
+            sg = ScenarioGraph()
+            sg._add_node('start')
+            self.assertIn('start', sg.networkx.nodes)
+            self.assertEqual(len(sg.networkx.nodes), 1)
 
-    def test_scenario_graph_update_visualisation_nodes(self):
-        ts = TraceState(3)
-        candidates = []
-        for scenario in range(3):
-            candidates.append(ts.next_candidate())
-            ts.confirm_full_scenario(candidates[-1], str(scenario), {})
-        ti = TraceInfo.from_trace_state(ts, ModelSpace())
-        sg = ScenarioGraph()
-        sg.update_visualisation(ti)
+        def test_scenario_graph_update_visualisation_nodes(self):
+            ts = TraceState(3)
+            candidates = []
+            for scenario in range(3):
+                candidates.append(ts.next_candidate())
+                ts.confirm_full_scenario(candidates[-1], str(scenario), {})
+            ti = TraceInfo.from_trace_state(ts, ModelSpace())
+            sg = ScenarioGraph()
+            sg.update_visualisation(ti)
 
-        self.assertIn('node0', sg.networkx.nodes)
-        self.assertEqual(sg.networkx.nodes['node0']['label'], str(0))
-        self.assertIn('node1', sg.networkx.nodes)
-        self.assertEqual(sg.networkx.nodes['node1']['label'], str(1))
-        self.assertIn('node2', sg.networkx.nodes)
-        self.assertEqual(sg.networkx.nodes['node2']['label'], str(2))
+            self.assertIn('node0', sg.networkx.nodes)
+            self.assertEqual(sg.networkx.nodes['node0']['label'], str(0))
+            self.assertIn('node1', sg.networkx.nodes)
+            self.assertEqual(sg.networkx.nodes['node1']['label'], str(1))
+            self.assertIn('node2', sg.networkx.nodes)
+            self.assertEqual(sg.networkx.nodes['node2']['label'], str(2))
 
-    def test_scenario_graph_update_visualisation_edges(self):
-        ts = TraceState(3)
-        candidates = []
-        for scenario in range(3):
-            candidates.append(ts.next_candidate())
-            ts.confirm_full_scenario(candidates[-1], str(scenario), {})
-        ti = TraceInfo.from_trace_state(trace=ts, state=ModelSpace())
-        sg = ScenarioGraph()
-        sg.update_visualisation(ti)
+        def test_scenario_graph_update_visualisation_edges(self):
+            ts = TraceState(3)
+            candidates = []
+            for scenario in range(3):
+                candidates.append(ts.next_candidate())
+                ts.confirm_full_scenario(candidates[-1], str(scenario), {})
+            ti = TraceInfo.from_trace_state(trace=ts, state=ModelSpace())
+            sg = ScenarioGraph()
+            sg.update_visualisation(ti)
 
-        self.assertIn(('node0', 'node1'), sg.networkx.edges)
-        self.assertIn(('node1', 'node2'), sg.networkx.edges)
+            self.assertIn(('node0', 'node1'), sg.networkx.edges)
+            self.assertIn(('node1', 'node2'), sg.networkx.edges)
 
-        edge_labels = nx.get_edge_attributes(sg.networkx, "label")
-        self.assertEqual(edge_labels[('node0', 'node1')], '')
-        self.assertEqual(edge_labels[('node1', 'node2')], '')
+            edge_labels = nx.get_edge_attributes(sg.networkx, "label")
+            self.assertEqual(edge_labels[('node0', 'node1')], '')
+            self.assertEqual(edge_labels[('node1', 'node2')], '')
 
-    def test_scenario_graph_update_visualisation_single_node(self):
-        ts = TraceState(1)
-        ts.confirm_full_scenario(0, 'one', {})
-        self.assertEqual(ts.get_trace(), ['one'])
-        ti = TraceInfo.from_trace_state(ts, ModelSpace())
-        sg = ScenarioGraph()
-        sg.update_visualisation(ti)
+        def test_scenario_graph_update_visualisation_single_node(self):
+            ts = TraceState(1)
+            ts.confirm_full_scenario(0, 'one', {})
+            self.assertEqual(ts.get_trace(), ['one'])
+            ti = TraceInfo.from_trace_state(ts, ModelSpace())
+            sg = ScenarioGraph()
+            sg.update_visualisation(ti)
 
-        # expected behaviour: no nodes nor edges are added
-        self.assertEqual(len(sg.networkx.nodes), 1)
-        self.assertEqual(len(sg.networkx.edges), 0)
+            # expected behaviour: no nodes nor edges are added
+            self.assertEqual(len(sg.networkx.nodes), 1)
+            self.assertEqual(len(sg.networkx.edges), 0)
 
-    def test_scenario_graph_set_starting_node_new_node(self):
-        sg = ScenarioGraph()
-        si = ScenarioInfo('test')
-        sg._set_starting_node(si)
-        node_id = sg._get_or_create_id(si)
-        # node
-        self.assertIn(node_id, sg.networkx.nodes)
-        self.assertEqual(sg.networkx.nodes[node_id]['label'], 'test')
+        def test_scenario_graph_set_starting_node_new_node(self):
+            sg = ScenarioGraph()
+            si = ScenarioInfo('test')
+            sg._set_starting_node(si)
+            node_id = sg._get_or_create_id(si)
+            # node
+            self.assertIn(node_id, sg.networkx.nodes)
+            self.assertEqual(sg.networkx.nodes[node_id]['label'], 'test')
 
-        # edge
-        self.assertIn(('start', node_id), sg.networkx.edges)
-        edge_labels = nx.get_edge_attributes(sg.networkx, "label")
-        self.assertEqual(edge_labels[('start', node_id)], '')
+            # edge
+            self.assertIn(('start', node_id), sg.networkx.edges)
+            edge_labels = nx.get_edge_attributes(sg.networkx, "label")
+            self.assertEqual(edge_labels[('start', node_id)], '')
 
-    def test_scenario_graph_set_starting_node_existing_node(self):
-        sg = ScenarioGraph()
-        si = ScenarioInfo('test')
-        node_id = sg._get_or_create_id(si)
-        sg._add_node(node_id)
-        self.assertIn(node_id, sg.networkx.nodes)
+        def test_scenario_graph_set_starting_node_existing_node(self):
+            sg = ScenarioGraph()
+            si = ScenarioInfo('test')
+            node_id = sg._get_or_create_id(si)
+            sg._add_node(node_id)
+            self.assertIn(node_id, sg.networkx.nodes)
 
-        sg._set_starting_node(si)
-        self.assertIn(('start', node_id), sg.networkx.edges)
-        edge_labels = nx.get_edge_attributes(sg.networkx, "label")
-        self.assertEqual(edge_labels[('start', node_id)], '')
+            sg._set_starting_node(si)
+            self.assertIn(('start', node_id), sg.networkx.edges)
+            edge_labels = nx.get_edge_attributes(sg.networkx, "label")
+            self.assertEqual(edge_labels[('start', node_id)], '')
 
-    def test_scenario_graph_set_end_node(self):
-        sg = ScenarioGraph()
-        si = ScenarioInfo('test')
-        node_id = sg._get_or_create_id(si)
-        sg._set_ending_node(si)
-        self.assertEqual(sg.end_node, node_id)
+        def test_scenario_graph_set_end_node(self):
+            sg = ScenarioGraph()
+            si = ScenarioInfo('test')
+            node_id = sg._get_or_create_id(si)
+            sg._set_ending_node(si)
+            self.assertEqual(sg.end_node, node_id)
 
 
 if __name__ == '__main__':
